Update topoGrid.C
The changes implemented try to make the deformation less dependent from the vertical extent of the original grid. Now the extent of the deformation is related to the horizontal extent, and the zero horizontal deformation is assigned at z=0.5*Ldef.
Above this level, the grid is still deformed, with the horizontal deformation going to zero lineary between 0.5*Ldef and zMax.