changes in Granular Viscosity Models
Now there are two models:
- Lun
- Gidaspow

The mono-disperse and poly-disperse for both the models should be consistent. Lun for the poly-disperse case correspond to the implementation of MFIX, where some corrections are present with respect to the original model. The mono-disperse Gidaspow model corresponds to that originally implemented in OpenFOAM (from van Wachem, Table 3.2, p.47). The poly-disperse formulation is based on the way Lun is generalized for the poly-disperse case in MFIX.